refactor: seperately move city hall page component
