+++ src/components/pages/city-hall/articles/ArticleRight.jsx
@@ -1,6 +1,6 @@
 import { HiInformationCircle } from "react-icons/hi";
 import HighlightTitle from "../ui/HighlightTitle";
-import BacoorCityHall from "../../../assets/images/bacoor-city-hall.png";
+import BacoorCityHall from "../../../../assets/images/bacoor-city-hall.png";
 
 const ArticleRight = () => {
   return (
